Make our email autoresponder more usable.
So far, our autoresponder would only send you bridges if your request
was valid.  Unfortunately, it's not very easy to figure out what a
correct request looks like.  This patch returns bridges regardless of if
the request was valid or not.  We also remove the "help" autoresponse
because there's no longer a need for it, and we simplify BridgeDB's auto
response, hopefully making it less frustrating for users.

This fixes <https://bugs.torproject.org/30941>.