Give BridgeDB more time to start before unit test.
The renaming of test_Bridges.py to test_bridgerings.py means that
test_bridgedb_script.py is now run earlier (because tests are executed
in alphabetic order), which means less time for BridgeDB to start.  This
commit adds an extra sleep to a unit test, so we can be reasonably sure
that BridgeDB is already running when unlinking its assignments file.