Only process text/plain part of emails.
When our autoresponder receives a multipart email from a user, it may
get confused by the parts that are not text/plain.  Instead of bending
over backwards to parse all sorts email encodings, this patch discard
the parts of a multipart email that aren't text/plain.

This fixes tpo/anti-censorship/bridgedb#33835.