Bug 40006: Parse remaining descs on validate error
If there was a faulty descriptor in the extrainfo document, Stem will
throw an exception and we loose the rest of the extrainfo descriptors
in the document. This passes extrainfo file objects to Stem so that if
an exception is raised on a single descriptor, the rest of the file can
be parsed starting at the file seek position.